chore: refactor signature types to different components (#380)
* chore: refactor signature types to different components

* fix: injection order of components
