+++ src/components/signatures/signTypedDataV3-sign.js
@@ -1,0 +1,164 @@
+import { recoverTypedSignature } from '@metamask/eth-sig-util';
+import { toChecksumAddress } from 'ethereumjs-util';
+import globalContext from '../..';
+import { EIP712Domain } from '../../signatures/utils';
+
+export function signTypedDataV3Component(parentContainer) {
+  parentContainer.insertAdjacentHTML(
+    'beforeend',
+    `<div class="col-xl-4 col-lg-6 col-md-12 col-sm-12 col-12 d-flex align-items-stretch">
+      <div class="card full-width">
+        <div class="card-body">
+          <h4>
+            Sign Typed Data V3
+          </h4>
+
+          <button
+            class="btn btn-primary btn-lg btn-block mb-3"
+            id="signTypedDataV3"
+            disabled
+          >
+            Sign
+          </button>
+
+          <p class="info-text alert alert-warning">
+            Result:
+            <span id="signTypedDataV3Result"></span>
+          </p>
+
+          <button
+            class="btn btn-primary btn-lg btn-block mb-3"
+            id="signTypedDataV3Verify"
+            disabled
+          >
+            Verify
+          </button>
+
+          <p class="info-text alert alert-warning">
+            Recovery result:
+            <span id="signTypedDataV3VerifyResult"></span>
+          </p>
+        </div>
+      </div>
+    </div>`,
+  );
+
+  const signTypedDataV3 = document.getElementById('signTypedDataV3');
+  const signTypedDataV3Result = document.getElementById(
+    'signTypedDataV3Result',
+  );
+  const signTypedDataV3Verify = document.getElementById(
+    'signTypedDataV3Verify',
+  );
+  const signTypedDataV3VerifyResult = document.getElementById(
+    'signTypedDataV3VerifyResult',
+  );
+
+  /**
+   * Sign Typed Data Version 3 Test
+   */
+  signTypedDataV3.onclick = async () => {
+    const msgParams = {
+      types: {
+        EIP712Domain,
+        Person: [
+          { name: 'name', type: 'string' },
+          { name: 'wallet', type: 'address' },
+        ],
+        Mail: [
+          { name: 'from', type: 'Person' },
+          { name: 'to', type: 'Person' },
+          { name: 'contents', type: 'string' },
+        ],
+      },
+      primaryType: 'Mail',
+      domain: {
+        name: 'Ether Mail',
+        version: '1',
+        chainId: globalContext.chainIdInt,
+        verifyingContract: '0xCcCCccccCCCCcCCCCCCcCcCccCcCCCcCcccccccC',
+      },
+      message: {
+        from: {
+          name: 'Cow',
+          wallet: '0xCD2a3d9F938E13CD947Ec05AbC7FE734Df8DD826',
+        },
+        to: {
+          name: 'Bob',
+          wallet: '0xbBbBBBBbbBBBbbbBbbBbbbbBBbBbbbbBbBbbBBbB',
+        },
+        contents: 'Hello, Bob!',
+      },
+    };
+    try {
+      const from = globalContext.accounts[0];
+      const sign = await globalContext.provider.request({
+        method: 'eth_signTypedData_v3',
+        params: [from, JSON.stringify(msgParams)],
+      });
+      signTypedDataV3Result.innerHTML = sign;
+      signTypedDataV3Verify.disabled = false;
+    } catch (err) {
+      console.error(err);
+      signTypedDataV3Result.innerHTML = `Error: ${err.message}`;
+    }
+  };
+
+  /**
+   * Sign Typed Data V3 Verification
+   */
+  signTypedDataV3Verify.onclick = async () => {
+    const msgParams = {
+      types: {
+        EIP712Domain,
+        Person: [
+          { name: 'name', type: 'string' },
+          { name: 'wallet', type: 'address' },
+        ],
+        Mail: [
+          { name: 'from', type: 'Person' },
+          { name: 'to', type: 'Person' },
+          { name: 'contents', type: 'string' },
+        ],
+      },
+      primaryType: 'Mail',
+      domain: {
+        name: 'Ether Mail',
+        version: '1',
+        chainId: globalContext.chainIdInt,
+        verifyingContract: '0xCcCCccccCCCCcCCCCCCcCcCccCcCCCcCcccccccC',
+      },
+      message: {
+        from: {
+          name: 'Cow',
+          wallet: '0xCD2a3d9F938E13CD947Ec05AbC7FE734Df8DD826',
+        },
+        to: {
+          name: 'Bob',
+          wallet: '0xbBbBBBBbbBBBbbbBbbBbbbbBBbBbbbbBbBbbBBbB',
+        },
+        contents: 'Hello, Bob!',
+      },
+    };
+    try {
+      const from = globalContext.accounts[0];
+      const sign = signTypedDataV3Result.innerHTML;
+      const recoveredAddr = await recoverTypedSignature({
+        data: msgParams,
+        signature: sign,
+        version: 'V3',
+      });
+      if (toChecksumAddress(recoveredAddr) === toChecksumAddress(from)) {
+        console.log(`Successfully verified signer as ${recoveredAddr}`);
+        signTypedDataV3VerifyResult.innerHTML = recoveredAddr;
+      } else {
+        console.log(
+          `Failed to verify signer when comparing ${recoveredAddr} to ${from}`,
+        );
+      }
+    } catch (err) {
+      console.error(err);
+      signTypedDataV3VerifyResult.innerHTML = `Error: ${err.message}`;
+    }
+  };
+}
