+++ src/signatures/utils.js
@@ -14,6 +14,16 @@ export const MSG_PRIMARY_TYPE = {
   SEAPORT_BULK_ORDER: 'BulkOrder',
 };
 
+export function splitSig(sig) {
+  const pureSig = sig.replace('0x', '');
+
+  const _r = Buffer.from(pureSig.substring(0, 64), 'hex');
+  const _s = Buffer.from(pureSig.substring(64, 128), 'hex');
+  const _v = Buffer.from(parseInt(pureSig.substring(128, 130), 16).toString());
+
+  return { _r, _s, _v };
+}
+
 export function getPermitMsgParams(
   { chainId, primaryType },
   { fromAddress } = {},
